feat: Añadir soporte para fechas en formato y registrar el nombre del producto en el modelo de oferta, operatividad de automatizacion de Cron

+++ apps/offers-ms/src/app/offers/offers.controller.ts
@@ -1,4 +1,4 @@
-import { BadRequestException, Controller } from '@nestjs/common';
+import { BadRequestException, Controller, Logger } from '@nestjs/common';
 import { MessagePattern, Payload } from '@nestjs/microservices';
 import { OffersService } from './offers.service';
 import { CreateOfferDto } from './dto/create-offer.dto';
@@ -6,18 +6,34 @@ import { UpdateOfferDto } from './dto/update-offer.dto';
 
 @Controller()
 export class OffersController {
+  private readonly logger = new Logger(OffersController.name);
+
   constructor(private readonly offersService: OffersService) {}
 
+  /**
+   * Crea una nueva oferta.
+   * @param createOfferDto 
+   */
   @MessagePattern('create_offer')
-  async create(@Payload() createOfferDto: CreateOfferDto) {
-    return await this.offersService.create(createOfferDto);
+  async handleCreateOffer(@Payload() createOfferDto: CreateOfferDto) {
+    this.logger.log(`Received create_offer pattern for productId: ${createOfferDto.productId}`);
+    return this.offersService.create(createOfferDto);
   }
+  
 
+  /**
+   * Obtiene todas las ofertas.
+   */
   @MessagePattern('get_all_offers')
   async findAll() {
     return await this.offersService.findAll();
   }
+    
 
+  /**
+   * Obtiene una oferta por su ID.
+   * @param payload 
+   */
   @MessagePattern('get_offer_by_id')
   async findOne(@Payload() payload: { id: string }) {
     const { id } = payload;
@@ -28,21 +44,33 @@ export class OffersController {
 
     return await this.offersService.findOne(id);
   }
-
+ 
+  
   @MessagePattern('update_offer')
   async update(@Payload() payload: { id: string; dto: UpdateOfferDto }) {
     const { id, dto } = payload;
+
+    if (!id || typeof id !== 'string') {
+      throw new BadRequestException('El ID proporcionado no es válido.');
+    }
+
+    if (!dto) {
+      throw new BadRequestException('El DTO de actualización es obligatorio.');
+    }
+
     return await this.offersService.update(id, dto);
   }
 
-/**
- * Elimina una oferta por su ID.
- * @param id - ID de la oferta.
- * @returns Una confirmación de eliminación.
- */
-@MessagePattern('delete_offer') // Patrón para eliminar una oferta
-async remove(@Payload() id: string) {
-  return await this.offersService.remove(id); // Devuelve la respuesta del servicio
-}
+  /**
+   * Elimina una oferta por su ID.
+   * @param id 
+  */
+  @MessagePattern('delete_offer')
+  async remove(@Payload() id: string) {
+    if (!id || typeof id !== 'string') {
+      throw new BadRequestException('El ID proporcionado no es válido.');
+    }
 
+    return await this.offersService.remove(id);
+  } 
 }
